+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 `StretchScrollView` adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.1](https://github.com/APUtils/StretchScrollView/releases/tag/3.0.1)
+Released on 09/25/2017.
+
+#### Fixed
+- Taking into account safe area adjusted offsets
+- Navigation bar background view fix
+
+
 ## [3.0.0](https://github.com/APUtils/StretchScrollView/releases/tag/3.0.0)
 Released on 09/21/2017.